*-without-instance-replacement: workaround to skip reboot on first run
The null_resource helper is decoupled from the Terraform state and
can't know whether it is ran the first time or not. On the first run
no reboot is needed because the instance just got created.
Keep the state in a local file to know whether the trigger ran already
or not. While having an old file around merely introduces an
unnecessary reboot when recreating an instance, deleting the file and
then making a userdata change is not supported, it has to be treated
like the terraform state file.