Restructure -without-instance-replacement variants to use flatcar-reset
With the selective OS reset provided by flatcar-reset in the latest
Alpha, we can drop the Ignition rootfs reformat workaround.
The terraform helper action for reprovisioning has been moved to a
script file and can now detect the first run to skip it as this would
mean to directly reprovision after provisioning (I see no good way to
express this in Terraform itself). For Azure SSH (also on a custom
port) or the az script command can be used for interaction. In the
future one could also think about having a service on the instance that
queries the userdata from the metadata API to detect changes.
Since this setup is used for a presentation, a VHS demo tape is also
brought in.